Bug 1577236 - clang-10: Fix a -Wimplicit-int-float-conversion warning in chromium r=bobowen
Fix:
/var/lib/jenkins/workspace/firefox-clang-lld-last/security/sandbox/chromium/base/time/time.h:796:18: error: implicit conversion from 'long' to 'double' changes value from 922337203
  return value > std::numeric_limits<int64_t>::max()
               ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Upstream code is very different now (uses some saturated_cast):
https://chromium.googlesource.com/chromium/src/base/+/master/time/time.h#846

Differential Revision: https://phabricator.services.mozilla.com/D43786

UltraBlame original commit: 03a3ed508dd820dd8d452b1c5af724bebe5a4b9f

+++ security/sandbox/chromium/base/time/time.h
@@ -793,7 +793,7 @@ constexpr TimeDelta TimeDelta::Min() {
 constexpr TimeDelta TimeDelta::FromDouble(double value) {
   
   
-  return value > std::numeric_limits<int64_t>::max()
+  return value > double(std::numeric_limits<int64_t>::max())
              ? Max()
              : value < std::numeric_limits<int64_t>::min()
                    ? Min()
